Created a grid class and Built some tools
Wrote a generic Grid class to be used for pathfinding.

Refactored the room prefabs which needed to derive from a child of the room prefab.

Since I didn't want to recreate the level geometry, I built some tools to convert the layout into a scriptable object so they could be regenerated using the new prefabs.  This SO could prove to be useful later down the road.

I needed to refactor some prefabs but didn't want to recreate the layout, so I built some tools to turn the geometry layout into a scriptable object.